refactor node naming on creation since it is no longer used for node renaming
